% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/insert2.R
\name{as.bst}
\alias{as.bst}
\title{Convert a named vector or list to a bst}
\usage{
as.bst(x, compare)
}
\arguments{
\item{x}{A named vector or list}

\item{compare}{The comparator function}
}
\description{
Convenience function to quickly populate a bst, given a named list/vector.
}

