% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bst.R
\name{bst}
\alias{bst}
\title{Create a new binary search tree}
\usage{
bst(keys = NULL, values = NULL, compare = compare_default)
}
\arguments{
\item{keys}{(optional) keys to initially insert into the tree}

\item{values}{(optional) values should be the same length as \code{keys}}

\item{compare}{The comparator that the tree should use for ordering}
}
\description{
Creates a new, empty, binary search tree. A \code{bst} stores keys and 
values. You can \link{insert} new key-value pairs in, and you can 
\link{retrieve} values out by providing a key.
}
\details{
The default comparator will use \code{<}, so is defined for 
any types with which you can use \code{<}.

Note that there is no checking for consistency of keys.
}
\examples{
mytree <- bst()
mytree
# add ("a": 3) to the tree
mytree2 <- insert(mytree, "a", 3)

# no side-effects
mytree
mytree2
retrieve(mytree, "a") ## returns NULL
retrieve(mytree2, "a") ## 3
}

