% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ceiling_of.R
\name{ceiling_of}
\alias{ceiling_of}
\title{Get the smallest key greater than or equal to a given key}
\usage{
ceiling_of(tree, key)
}
\arguments{
\item{tree}{A \code{bst}}

\item{key}{The key to use in comparisons}
}
\description{
Get the smallest key greater than or equal to a given key
}
\details{
If the key exists in the tree, then it is returned. If the key is
larger than the largest key in the tree, then \code{NULL} is returned.
Otherwise, the smallest key that is greater than the given key is returned.
}
\examples{
mytree <- bst()
mytree <- insert(mytree, 1, "val1")
mytree <- insert(mytree, 3, "val2")
retrieve(mytree, ceiling_of(mytree, 2))
retrieve(mytree, floor_of(mytree, 3))
}

