% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/compare.R
\name{compare}
\alias{compare}
\title{Compare two keys}
\usage{
compare(key1, key2)
}
\description{
To create a custom \code{compare} method for a class, define a method that
takes two objects of the class, \code{key1} and \code{key2}, and returns
-1, 0, or 1.
}

