% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/compare.R
\name{compare_default}
\alias{compare_default}
\title{Compare two keys}
\usage{
compare_default(key1, key2)
}
\arguments{
\item{key1}{Keys to compare}

\item{key2}{Keys to compare}
}
\description{
To use your own comparator, create a function that compares two keys and
returns an integer, and use it as an argument when constructing a
\code{bst}
}

