% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/contains.R
\name{contains}
\alias{contains}
\title{Does the tree contain a given key?}
\usage{
contains(tree, key)
}
\arguments{
\item{tree}{A \code{bst} object}

\item{key}{The key to search for}
}
\description{
Does the tree contain a given key?
}
\details{
returns TRUE or FALSE
}

