% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete.R
\name{delete}
\alias{delete}
\title{Delete a key-value pair from a tree}
\usage{
delete(tree, key)
}
\arguments{
\item{tree}{A \code{bst}}

\item{key}{The key to delete}
}
\description{
Delete a key-value pair from a tree
}
\details{
If the tree contains the key, then a new tree with the key removed
is returned. Otherwise, the original tree is returned, with a warning
}

