% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/floor_of.R
\name{floor_of}
\alias{floor_of}
\title{Get the largest key less than or equal to a given key}
\usage{
floor_of(tree, key)
}
\arguments{
\item{tree}{A \code{bst}}

\item{key}{The key to use in comparisons}
}
\description{
Get the largest key less than or equal to a given key
}
\details{
If the key exists in the tree, then it is returned. If the key is
smaller than the smallest key in the tree, then \code{NULL} is returned.
Otherwise, the largest key that is smaller than the given key is returned.
}
\examples{
mytree <- bst()
mytree <- insert(mytree, 1, "val1")
mytree <- insert(mytree, 3, "val2")
retrieve(mytree, floor_of(mytree, 2))
retrieve(mytree, floor_of(mytree, 4))
}

