% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/insert.R
\name{insert}
\alias{[<-.bst}
\alias{insert}
\title{Insert a key-value pair into a bst}
\usage{
insert(tree, key, value, ...)

\method{[}{bst}(tree, key) <- value
}
\arguments{
\item{tree}{A \code{bst}}

\item{key}{The key to insert}

\item{value}{The value to insert}

\item{...}{Other arguments (not used)}
}
\description{
If the given key is already present in the tree, the associated value will
be overwritten, with a warning.
}
\examples{
mytree <- bst()
mytree <- insert(mytree, "a", 7)

# there is no type-checking, so you can do, eg:
mytree <- insert(mytree, 3.14, 9)

# insert has no  side-effects
mytree2 <- insert(mytree, "k", 10)
retrieve(mytree, "k") # NULL
retrieve(mytree2, "k") # 10
# or use [
mytree["q"] <- 15
}

