% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minmax.R
\name{min_key}
\alias{max_key}
\alias{min_key}
\title{Get the smallest or largest key}
\usage{
min_key(tree)

max_key(tree)
}
\arguments{
\item{tree}{A \code{bst}}
}
\description{
Get the smallest or largest key
}
\section{Functions}{
\itemize{
\item \code{max_key}: Get the largest key
}}

