% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mockbst.R
\name{mockbst}
\alias{mockbst}
\title{Create a mock bst object for testing and analysis}
\usage{
mockbst(nodes, data_size = 1, key_type = "string")
}
\arguments{
\item{nodes}{Size of the tree}

\item{data_size}{Size of each data element}

\item{key_type}{"string" or "integer"}
}
\description{
Create a mock bst object for testing and analysis
}

