% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/prune.R
\name{prune}
\alias{prune}
\title{Return a subsetted tree}
\usage{
prune(tree, low, high)
}
\arguments{
\item{tree}{The \code{bst}}

\item{low}{Low end of the range to return}

\item{high}{high end of the range to return}
}
\description{
Prune returns a new \code{bst} with all keys (and associated values) between
low and high.
}
\details{
Returns a new \code{bst}
}

