% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/retrieve.R
\name{retrieve}
\alias{[.bst}
\alias{retrieve}
\title{Retrieve the value associated with a given key}
\usage{
retrieve(tree, key, ...)

\method{[}{bst}(tree, key)
}
\arguments{
\item{tree}{A \code{bst}}

\item{key}{The key to search for}
}
\description{
If the key exists in the tree, \code{retrieve} returns the associated value.
Otherwise, \code{retrieve} returns \code{NULL}.
}
\examples{
mytree <- bst()
mytree <- insert(mytree, "a", 7)
retrieve(mytree, "a")
# can also use subset operator
mytree["a"]
}

