% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/insert2.R
\name{as.bst}
\alias{as.bst}
\title{Convert a named vector or list to a bst}
\usage{
as.bst(x)
}
\description{
Convenience function to quickly populate a bst, given a named list/vector.
}
\details{
The normal \link{insert} operation has to make copies of some of
the nodes of a tree in order to maintain the non-destructive nature of the
update -- on average this will be around \code{log(n)} copies where n is
the size of the \code{bst}. In the special case where a new tree is created
and populated from an existing list, so there is no expectation of
continued access to versions of the tree between inserts, we save a bit of
time and unnecessary work by inserting without any copying.
}

